components/requirements.txt
tensorflow_addons==0.12.1
tensorflow_gpu==2.4.1
numpy==1.19.5
